Random letters after clicking arrow, saving settings, checking user answer
pick_georgian_letter() method was expanded. Now it chooses 4 random
letters in user native language and assigns them to answer buttons.
It's triggered when the arrow is pressed.

check_answer() method was created. For now it only check if user
answer is correc or not.

save_settings() method was created. It saves current settings
to the .csv file. It's triggered when user changes language.