Implemented achievements for Letter Learning mode.
Created check_achievement_letters method which checks if a letter
was already answered correctly. If not, it's added to a list and
number of correctly answered letters is increased. Upon treshold
achievements are activated.

After obtaining an achievement, appropriate tile is 'clickable'
in Achievements screen.

Status of achievements is stored the file, so it's not lost when
the app is closed.